refactor(providers): extract application bindings
